suggest::benchmarks::client

Struct RemoteSettingsBenchmarkClient

pub struct RemoteSettingsBenchmarkClient { /* private fields */ }

Remotes settings client for benchmarking

This fetches all data in new, then implements [rs::Client] by returning the local data. Construct this one before the benchmark is run, then clone it as input for the benchmark. This ensures that network time does not count towards the benchmark time.

Implementations

impl RemoteSettingsBenchmarkClient

pub fn new() -> Result<Self, Error>

pub fn attachment_size_by_record_type(&self) -> Vec<(SuggestRecordType, usize)>

pub fn total_attachment_size(&self) -> usize