Enum suggest::Suggestion

pub enum Suggestion {
    Amp {

        title: String,
        url: String,
        raw_url: String,
        icon: Option<Vec<u8>>,
        icon_mimetype: Option<String>,
        full_keyword: String,
        block_id: i64,
        advertiser: String,
        iab_category: String,
        impression_url: String,
        click_url: String,
        raw_click_url: String,
        score: f64,
    
},
    Pocket {
        title: String,
        url: String,
        score: f64,
        is_top_pick: bool,
    },
    Wikipedia {
        title: String,
        url: String,
        icon: Option<Vec<u8>>,
        icon_mimetype: Option<String>,
        full_keyword: String,
    },
    Amo {
        title: String,
        url: String,
        icon_url: String,
        description: String,
        rating: Option<String>,
        number_of_ratings: i64,
        guid: String,
        score: f64,
    },
    Yelp {
        url: String,
        title: String,
        icon: Option<Vec<u8>>,
        icon_mimetype: Option<String>,
        score: f64,
        has_location_sign: bool,
        subject_exact_match: bool,
        location_param: String,
    },
    Mdn {
        title: String,
        url: String,
        description: String,
        score: f64,
    },
    Weather {
        score: f64,
    },
    Fakespot {
        fakespot_grade: String,
        product_id: String,
        rating: f64,
        title: String,
        total_reviews: i64,
        url: String,
        icon: Option<Vec<u8>>,
        icon_mimetype: Option<String>,
        score: f64,
    },
    Exposure {
        suggestion_type: String,
        score: f64,
    },
}

A suggestion from the database to show in the address bar.

Variants

Amp

Fields

title: String

url: String

raw_url: String

icon: Option<Vec<u8>>

icon_mimetype: Option<String>

full_keyword: String

block_id: i64

advertiser: String

iab_category: String

impression_url: String

click_url: String

raw_click_url: String

score: f64

Pocket

Fields

title: String

url: String

score: f64

is_top_pick: bool

Wikipedia

Fields

title: String

url: String

icon: Option<Vec<u8>>

icon_mimetype: Option<String>

full_keyword: String

Amo

Fields

title: String

url: String

icon_url: String

description: String

rating: Option<String>

number_of_ratings: i64

guid: String

score: f64

Yelp

Fields

url: String

title: String

icon: Option<Vec<u8>>

icon_mimetype: Option<String>

score: f64

has_location_sign: bool

subject_exact_match: bool

location_param: String

Mdn

Fields

title: String

url: String

description: String

score: f64

Weather

Fields

score: f64

Fakespot

Fields

fakespot_grade: String

product_id: String

rating: f64

title: String

total_reviews: i64

url: String

icon: Option<Vec<u8>>

icon_mimetype: Option<String>

score: f64

Exposure

Fields

suggestion_type: String

score: f64

Implementations

impl Suggestion

pub fn url(&self) -> Option<&str>

Get the URL for this suggestion, if present

pub fn raw_url(&self) -> Option<&str>

Get the raw URL for this suggestion, if present

This is the same as url except for Amp. In that case, url is the URL after being “cooked” using template interpolation, while raw_url is the URL template.

pub fn title(&self) -> &str

pub fn icon_data(&self) -> Option<&[u8]>

pub fn score(&self) -> f64