Crate suggest

Modules

benchmarks

Benchmarking support

util

Structs

Geoname

A single geographic place.

GeonameMatch

A fetched geoname with info on how it was matched.

LabeledTimingSample

Single sample for a Glean labeled_timing_distribution

QueryWithMetricsResult

SuggestGlobalConfig

Global Suggest configuration data.

SuggestIngestionConstraints

Constraints limit which suggestions to ingest from Remote Settings.

SuggestIngestionMetrics

Ingestion metrics

SuggestStore

The store is the entry point to the Suggest component. It incrementally downloads suggestions from the Remote Settings service, stores them in a local database, and returns them in response to user queries.

SuggestStoreBuilder

Builder for SuggestStore

SuggestionProviderConstraints

Some providers manage multiple suggestion subtypes. Queries, ingests, and other operations on those providers must be constrained to a desired subtype.

SuggestionQuery

A query for suggestions to show in the address bar.

Enums

AmpMatchingStrategy

Error

A list of errors that are internal to the component. This is the error type for private and crate-internal methods, and is never returned to the application.

InterruptKind

What should be interrupted when SuggestStore::interrupt is called?

SuggestApiError

The error type for all Suggest component operations. These errors are exposed to your application, which should handle them as needed.

SuggestProviderConfig

Per-provider configuration data.

Suggestion

A suggestion from the database to show in the address bar.

SuggestionProvider

A provider is a source of search suggestions. Please preserve the integer values after removing or adding providers. Provider configs are associated with integer keys stored in the database.

Functions

raw_suggestion_url_matches

Determines whether a “raw” sponsored suggestion URL is equivalent to a “cooked” URL. The two URLs are equivalent if they are identical except for their replaced template parameters, which can be different.

Type Aliases

SuggestApiResult