suggest

Enum Suggestion

pub enum Suggestion {
    Amp {

        title: String,
        url: String,
        raw_url: String,
        icon: Option<Vec<u8>>,
        icon_mimetype: Option<String>,
        full_keyword: String,
        block_id: i64,
        advertiser: String,
        iab_category: String,
        categories: Vec<i32>,
        impression_url: String,
        click_url: String,
        raw_click_url: String,
        score: f64,
        fts_match_info: Option<FtsMatchInfo>,
    
},
    Wikipedia {
        title: String,
        url: String,
        icon: Option<Vec<u8>>,
        icon_mimetype: Option<String>,
        full_keyword: String,
    },
    Amo {
        title: String,
        url: String,
        icon_url: String,
        description: String,
        rating: Option<String>,
        number_of_ratings: i64,
        guid: String,
        score: f64,
    },
    Yelp {
        url: String,
        title: String,
        icon: Option<Vec<u8>>,
        icon_mimetype: Option<String>,
        score: f64,
        has_location_sign: bool,
        subject_exact_match: bool,
        subject_type: YelpSubjectType,
        location_param: String,
    },
    Mdn {
        title: String,
        url: String,
        description: String,
        score: f64,
    },
    Weather {
        city: Option<Geoname>,
        score: f64,
    },
    Fakespot {
        fakespot_grade: String,
        product_id: String,
        rating: f64,
        title: String,
        total_reviews: i64,
        url: String,
        icon: Option<Vec<u8>>,
        icon_mimetype: Option<String>,
        score: f64,
        match_info: Option<FtsMatchInfo>,
    },
    Dynamic {
        suggestion_type: String,
        data: Option<Value>,
        dismissal_key: Option<String>,
        score: f64,
    },
}

A suggestion from the database to show in the address bar.

Variants

Amp

Fields

title: String

url: String

raw_url: String

icon: Option<Vec<u8>>

icon_mimetype: Option<String>

full_keyword: String

block_id: i64

advertiser: String

iab_category: String

categories: Vec<i32>

impression_url: String

click_url: String

raw_click_url: String

score: f64

fts_match_info: Option<FtsMatchInfo>

Wikipedia

Fields

title: String

url: String

icon: Option<Vec<u8>>

icon_mimetype: Option<String>

full_keyword: String

Amo

Fields

title: String

url: String

icon_url: String

description: String

rating: Option<String>

number_of_ratings: i64

guid: String

score: f64

Yelp

Fields

url: String

title: String

icon: Option<Vec<u8>>

icon_mimetype: Option<String>

score: f64

has_location_sign: bool

subject_exact_match: bool

subject_type: YelpSubjectType

location_param: String

Mdn

Fields

title: String

url: String

description: String

score: f64

Weather

Fields

city: Option<Geoname>

score: f64

Fakespot

Fields

fakespot_grade: String

product_id: String

rating: f64

title: String

total_reviews: i64

url: String

icon: Option<Vec<u8>>

icon_mimetype: Option<String>

score: f64

match_info: Option<FtsMatchInfo>

Dynamic

Fields

suggestion_type: String

data: Option<Value>

dismissal_key: Option<String>

This value is optionally defined in the suggestion’s remote settings data and is an opaque token used for dismissing the suggestion in lieu of a URL. If Some, the suggestion can be dismissed by passing the wrapped string to crate::SuggestStore::dismiss_suggestion.

score: f64

Implementations

impl Suggestion

pub fn dismissal_key(&self) -> Option<&str>

Get the suggestion’s dismissal key, which should be stored in the dismissed_suggestions table when the suggestion is dismissed. Some suggestions may not have dismissal keys and cannot be dismissed.

pub fn url(&self) -> Option<&str>

Get the URL for this suggestion, if present

pub fn raw_url(&self) -> Option<&str>

Get the raw URL for this suggestion, if present

This is the same as url except for Amp. In that case, url is the URL after being “cooked” using template interpolation, while raw_url is the URL template.

pub fn title(&self) -> &str

pub fn icon_data(&self) -> Option<&[u8]>

pub fn score(&self) -> f64

pub fn fts_match_info(&self) -> Option<&FtsMatchInfo>